Migrate project to .NET 5
