--- Patients/packages.config
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="EntityFramework" version="6.4.0" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="12.0.1" targetFramework="net461" />
-  <package id="SQLite.CodeFirst" version="1.6.0.30" targetFramework="net461" />
-  <package id="StyleCop.Analyzers" version="1.1.118" targetFramework="net461" developmentDependency="true" />
-  <package id="System.Data.SQLite" version="1.0.109.2" targetFramework="net461" />
-  <package id="System.Data.SQLite.Core" version="1.0.109.2" targetFramework="net461" />
-  <package id="System.Data.SQLite.EF6" version="1.0.109.0" targetFramework="net461" />
-  <package id="System.Data.SQLite.Linq" version="1.0.109.0" targetFramework="net461" />
-</packages>
